BMS project

+++ src/main/java/com/corebanking/engine/infrastructure/web/dto/request/OpenAccountRequest.java
@@ -1,9 +1,9 @@
 package com.corebanking.engine.infrastructure.web.dto.request;
 
-public record OpenAccountRequest
-(
-    String customerId,
-    String accountType,
-    double initialDeposit
-) 
-{}
+import java.math.BigDecimal;
+
+public record OpenAccountRequest(
+        String customerId,
+        String accountType,
+        BigDecimal initialDeposit
+) {}
